implemented thumbnail, navigation, and auto update via swr

+++ backend/src/draw.rs
@@ -3,6 +3,6 @@ use crate::schema::*;
 #[derive(Insertable)]
 #[table_name = "draws"]
 pub struct NewDraw {
-    pub room_id: i32,
-    pub path_id: i32,
+    pub room_id: i64,
+    pub path_id: i64,
 }
